+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.11.0
 - FF-1429 - Updated NuGet to 5.11.0
 - FF-1429 - Updated Roslynator.Analyzers to 3.2.2
+- FF-1429 - Updated FunFair.CodeAnalysis to 5.4.0.915
 ### Removed
 ### Deployment Changes