+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated FunFair.CodeAnalysis to 5.6.0.960
+- FF-1429 - Updated FunFair.Test.Common to 5.6.1.1333
 ### Removed
 ### Deployment Changes