+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Changed
 - FF-1429 - Updated FunFair.CodeAnalysis to 5.6.0.960
 - FF-1429 - Updated FunFair.Test.Common to 5.6.1.1333
+- FF-1429 - Updated FunFair.Test.Common to 5.6.2.1337
 ### Removed
 ### Deployment Changes